\name{as.big.nifti4d-methods}
\alias{as.big.nifti4d-methods}
\alias{as.big.nifti4d-methods}
\title{Create a big...}
\description{Create a big.nifti4d object}
\alias{as.big.nifti4d,nifti-method}
\alias{as.big.nifti4d,nifti4d-method}
\alias{as.big.nifti4d,array,list-method}
\alias{as.big.nifti4d,matrix,list,logical-method}
\alias{as.big.nifti4d,big.matrix,list,logical-method}
\section{Methods}{\describe{
\item{\code{signature(x="nifti")}}{...}
\item{\code{signature(x="nifti4d")}}{...}
\item{\code{signature(x="array", header="list")}}{...}
\item{\code{signature(x="matrix", header="list", mask="logical")}}{...}
\item{\code{signature(x="big.matrix", header="list", mask="logical")}}{...}
}

}
\seealso{\code{\link{as.big.nifti4d}}}
\keyword{methods}

