\name{as.nifti3d-methods}
\alias{as.nifti3d-methods}
\alias{as.nifti3d-methods}
\title{Create a nifti3d object...}
\description{Create a nifti3d object}
\alias{as.nifti3d,nifti-method}
\alias{as.nifti3d,array,list-method}
\alias{as.nifti3d,vector,list,logical-method}
\section{Methods}{\describe{
\item{\code{signature(x="nifti")}}{...}
\item{\code{signature(x="array", header="list")}}{...}
\item{\code{signature(x="vector", header="list", mask="logical")}}{...}
}

}
\seealso{\code{\link{as.nifti3d}}}
\keyword{methods}

