\name{as.nifti4d-methods}
\alias{as.nifti4d-methods}
\alias{as.nifti4d-methods}
\title{Create a nifti4d object...}
\description{Create a nifti4d object}
\alias{as.nifti4d,nifti-method}
\alias{as.nifti4d,big.nifti4d-method}
\alias{as.nifti4d,array,list-method}
\alias{as.nifti4d,matrix,list,logical-method}
\section{Methods}{\describe{
\item{\code{signature(x="nifti")}}{...}
\item{\code{signature(x="big.nifti4d")}}{...}
\item{\code{signature(x="array", header="list")}}{...}
\item{\code{signature(x="matrix", header="list", mask="logical")}}{...}
}

}
\seealso{\code{\link{as.nifti4d}}}
\keyword{methods}

