\name{nifti-class}
\alias{nifti-class}
\alias{nifti}
\title{nifti/analyze class}
\description{Class returned when reading a nifti/analyze file or from \code{\link{nifti}}/\code{\link{as.nifti}}}
\section{Slots}{\describe{\item{\code{header}:}{}\item{\code{.Data}:}{}}}
\section{Extends}{\code{\linkS4class{header}}, \code{\linkS4class{array}}}

