\name{nifti3d-class}
\alias{nifti3d-class}
\alias{nifti3d}
\title{3d nifti => 1d vector}
\description{Represents a 3D nifti image as a 1D vector}
\section{Slots}{\describe{\item{\code{header}:}{}\item{\code{mask}:}{}\item{\code{.Data}:}{}}}
\section{Extends}{\code{\linkS4class{header}}, \code{\linkS4class{niftiXd}}, \code{\linkS4class{vector}}}

