\name{niftir-package}
\alias{niftir-package}
\alias{niftir-package}
\title{Read/write nifti/analyze files}
\description{Read/write nifti/analyze files}
\details{Do Stuff

\tabular{ll}{
Package: \tab niftir\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2010-08-24\cr
License: \tab LGPL-3\cr
Copyright: \tab (C) 2010 Zarrar Shehzad and John W. Emerson\cr
LazyLoad: \tab yes\cr
}}
\alias{niftir}
\docType{package}
\author{Zarrar Shehzad}
\keyword{package}

