\name{niftiXd-class}
\alias{niftiXd-class}
\alias{niftiXd}
\title{Not really used on its own (but the class has friends that extend it)...}
\description{Not really used on its own (but the class has friends that extend it)}
\section{Slots}{\describe{\item{\code{mask}:}{(\code{\link{vector}}) A logical vector (represents the voxels in the original 3D image whose values are given in \code{.Data} or \code{address}); its length must equal the product of the dim attribute in the header}}}

