\name{read.nifti.header}
\alias{read.nifti.header}
\title{Read header info from nifti/analyze file...}
\usage{read.nifti.header(fname)
}
\description{Read header info from nifti/analyze file}
\author{Zarrar Shehzad}
\value{list with header attributes
(see output of \code{\link{create.header}} for supported attributes)}
\note{Some of the C code that this function calls was borrowed
from Rniftilib}
\seealso{\code{\link{read.nifti}}}
\arguments{\item{fname}{nifti/analyze filename}
}

