\name{read.nifti4d}
\alias{read.nifti4d}
\title{Read in a nifti4d object from a file...}
\usage{read.nifti4d(fname)}
\description{Read in a nifti4d object from a file}
\author{Zarrar Shehzad}
\value{\code{nifti4d} object}
\seealso{\code{\link{read.nifti}}, \code{\link{as.nifti4d}}}
\keyword{methods}
\arguments{\item{fname}{character specifying path to analyze/nifti file}
}
\examples{# TODO}

