\name{write.nifti-methods}
\alias{write.nifti-methods}
\alias{write.nifti-methods}
\title{Write nifti/analyze file...}
\description{Write nifti/analyze file}
\alias{write.nifti,nifti-method}
\alias{write.nifti,big.nifti4d-method}
\alias{write.nifti,ANY-method}
\section{Methods}{\describe{
\item{\code{signature(x="nifti")}}{...}
\item{\code{signature(x="big.nifti4d")}}{...}
\item{\code{signature(x="ANY")}}{...}
}

}
\seealso{\code{\link{write.nifti}}}
\keyword{methods}

